refactor: split auto_authn orm tables
