+++ pkgs/standards/auto_authn/auto_authn/v2/orm/__init__.py
@@ -1,0 +1,3 @@
+"""ORM table models for auto_authn."""
+
+from .tables import *  # noqa: F401,F403
